[test] viewDidLoadWithoutAnswers_shouldNot_renderAnswers
